chore: Refactors Reanimated screen to fix lint error
