+++ src/reanimated/index.tsx
@@ -1,2 +1,2 @@
-export { default as ReanimatedScreenProvider } from './ReanimatedScreen';
+export { default as ReanimatedScreenProvider } from './ReanimatedScreenProvider';
 export { default as useReanimatedTransitionProgress } from './useReanimatedTransitionProgress';
